feat: create example build serve
