--- example/README.md
@@ -1,2 +1,0 @@
-# vue-pinch-scroll-zoom
-Vue component that provides content scrolling and zooming using mouse events or two fingers pinch on a mobile devices
