--- example/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@coddicat/vue-pinch-scroll-zoom-example",
-  "version": "3.3.1",
-  "description": "Vue component that provides content scrolling and zooming using mouse events or two fingers pinch on a mobile devices",
-  "author": "Rodion Shlomo Solomonyk",
-  "homepage": "https://github.com/coddicat/vue-pinch-scroll-zoom",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "vue": "^3.0.5",
-    "core-js": "^3.6.5"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "typescript": "~4.1.5"
-  }
-}
