--- example/src/main.ts
@@ -1,6 +1,0 @@
-import { createApp, h } from 'vue'
-import App from './App.vue'
-
-const app = createApp(App)
-
-app.mount('#app')
