--- example/src/shims-vue.d.ts
@@ -1,6 +1,0 @@
-declare module '*.vue' {
-  import { DefineComponent } from 'vue';
-
-  const Component: DefineComponent<{}, {}, any>;
-  export default Component;
-}
