+++ package.json
@@ -61,8 +61,5 @@
     "ttypescript": "^1.5.12",
     "typescript": "^4.0.3",
     "vue": "^3.0.5"
-  },
-  "dependencies": {
-    "webpack": "^5.72.0"
   }
 }
